Add files via upload
The Hospital Management System is developed to improve the efficiency of hospital tasks and operations by organization and management of data through a database. The system handles day-to-day tasks occurring inside a hospital such as patient care, staff management, room allocation, appointment bookings, billing, admissions and medical records.
In many healthcare settings, and especially larger organizations, manual record-keeping leads to errors, delays, and data loss. This system addresses those challenges by providing a reliable platform for storing, updating, and retrieving important hospital related data.
A brief over the applications used to design this software is that we used PHP in line with HTML docs. The integration and deployment of the project are managed through XAMPP.